+++ events/ready.js
@@ -1,4 +1,6 @@
 const weekday = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
+const authorize = require("../functions/googleAuthentication.js");
+const { google } = require('googleapis');
 const cron = require("cron");
 const fs = require('fs');
 
@@ -12,79 +14,13 @@ module.exports = client => {
     client.guild = guild;
 
 	let scheduleCheck = new cron.CronJob('00 59 5,11,17,23 * * *', () => {
-        const readline = require('readline');
-        const { google } = require('googleapis');
 
-        /* Start of Google Stuff */
-        //
-        //
-        //
-        // If modifying these scopes, delete token.json.
-        const SCOPES = ['https://www.googleapis.com/auth/calendar.readonly'];
-        // The file token.json stores the user's access and refresh tokens, and is
-        // created automatically when the authorization flow completes for the first
-        // time.
-        const TOKEN_PATH = 'token.json';
-
-        // Load client secrets from a local file.
         fs.readFile('../credentials.json', (err, content) => {
             if (err) return console.log('Error loading client secret file:', err);
             // Authorize a client with credentials, then call the Google Calendar API.
             authorize(JSON.parse(content), checkEvents);
         });
 
-        /**
-         * Create an OAuth2 client with the given credentials, and then execute the
-         * given callback function.
-         * @param {Object} credentials The authorization client credentials.
-         * @param {function} callback The callback to call with the authorized client.
-         */
-        function authorize(credentials, callback) {
-            const { client_secret, client_id, redirect_uris } = credentials.installed;
-            const oAuth2Client = new google.auth.OAuth2(
-                client_id, client_secret, redirect_uris[0]);
-
-            // Check if we have previously stored a token.
-            fs.readFile(TOKEN_PATH, (err, token) => {
-                if (err) return getAccessToken(oAuth2Client, callback);
-                oAuth2Client.setCredentials(JSON.parse(token));
-                callback(oAuth2Client);
-            });
-        }
-
-        /**
-         * Get and store new token after prompting for user authorization, and then
-         * execute the given callback with the authorized OAuth2 client.
-         * @param {google.auth.OAuth2} oAuth2Client The OAuth2 client to get token for.
-         * @param {getEventsCallback} callback The callback for the authorized client.
-         */
-        function getAccessToken(oAuth2Client, callback) {
-            const authUrl = oAuth2Client.generateAuthUrl({
-                access_type: 'offline',
-                scope: SCOPES,
-            });
-            console.log('Authorize this app by visiting this url:', authUrl);
-            const rl = readline.createInterface({
-                input: process.stdin,
-                output: process.stdout,
-            });
-            rl.question('Enter the code from that page here: ', (code) => {
-                rl.close();
-                oAuth2Client.getToken(code, (err, token) => {
-                    if (err) return console.error('Error retrieving access token', err);
-                    oAuth2Client.setCredentials(token);
-                    // Store the token to disk for later program executions
-                    fs.writeFile(TOKEN_PATH, JSON.stringify(token), (err) => {
-                        if (err) return console.error(err);
-                        console.log('Token stored to', TOKEN_PATH);
-                    });
-                    callback(oAuth2Client);
-                });
-            });
-        }
-        //
-        //
-        //
         /* End of Google Stuff */
 
         function checkEvents(auth) {
@@ -138,24 +74,12 @@ module.exports = client => {
             }
         }
 
-        const readline = require('readline');
-        const { google } = require('googleapis');
-
         const eventDates = [];
         const eventSummaries = [];
         const eventDescriptions = [];
         const eventColors = [];
 
         /* Start of Google Stuff */
-        //
-        //
-        //
-        // If modifying these scopes, delete token.json.
-        const SCOPES = ['https://www.googleapis.com/auth/calendar.readonly'];
-        // The file token.json stores the user's access and refresh tokens, and is
-        // created automatically when the authorization flow completes for the first
-        // time.
-        const TOKEN_PATH = 'token.json';
 
         // Load client secrets from a local file.
         fs.readFile('../credentials.json', (err, content) => {
@@ -163,59 +87,6 @@ module.exports = client => {
             // Authorize a client with credentials, then call the Google Calendar API.
             authorize(JSON.parse(content), weekEvents);
         });
-
-        /**
-         * Create an OAuth2 client with the given credentials, and then execute the
-         * given callback function.
-         * @param {Object} credentials The authorization client credentials.
-         * @param {function} callback The callback to call with the authorized client.
-         */
-        function authorize(credentials, callback) {
-            const { client_secret, client_id, redirect_uris } = credentials.installed;
-            const oAuth2Client = new google.auth.OAuth2(
-                client_id, client_secret, redirect_uris[0]);
-
-            // Check if we have previously stored a token.
-            fs.readFile(TOKEN_PATH, (err, token) => {
-                if (err) return getAccessToken(oAuth2Client, callback);
-                oAuth2Client.setCredentials(JSON.parse(token));
-                callback(oAuth2Client);
-            });
-        }
-
-        /**
-         * Get and store new token after prompting for user authorization, and then
-         * execute the given callback with the authorized OAuth2 client.
-         * @param {google.auth.OAuth2} oAuth2Client The OAuth2 client to get token for.
-         * @param {getEventsCallback} callback The callback for the authorized client.
-         */
-        function getAccessToken(oAuth2Client, callback) {
-            const authUrl = oAuth2Client.generateAuthUrl({
-                access_type: 'offline',
-                scope: SCOPES,
-            });
-            console.log('Authorize this app by visiting this url:', authUrl);
-            const rl = readline.createInterface({
-                input: process.stdin,
-                output: process.stdout,
-            });
-            rl.question('Enter the code from that page here: ', (code) => {
-                rl.close();
-                oAuth2Client.getToken(code, (err, token) => {
-                    if (err) return console.error('Error retrieving access token', err);
-                    oAuth2Client.setCredentials(token);
-                    // Store the token to disk for later program executions
-                    fs.writeFile(TOKEN_PATH, JSON.stringify(token), (err) => {
-                        if (err) return console.error(err);
-                        console.log('Token stored to', TOKEN_PATH);
-                    });
-                    callback(oAuth2Client);
-                });
-            });
-        }
-        //
-        //
-        //
         /* End of Google Stuff */
